rearange meili related files
