+++ next/backend/meili/types.ts
@@ -1,4 +1,18 @@
-import { BlogPost, Page, PageCategory, Tag, UploadFile } from '../graphql'
+import { BlogPost, Page, PageCategory, Tag, UploadFile, Vzn } from '../../graphql'
+
+/**
+ * A type that describes an entity wrapped in shared search index.
+ * E.g.:
+ * {
+ *     type: "article",
+ *     article: {...}
+ * }
+ */
+export type SearchIndexWrapped<T extends string, K extends object> = {
+  [key in T]: K
+} & {
+  type: T
+}
 
 // Meilisearch doesn't nest entities in `data.attributes`, therefore in order to use Strapi types we need to `Omit` those
 // attributes that are nested and replace them with their direct representations.
@@ -16,4 +30,12 @@ export type BlogPostMeili = Omit<BlogPost, '__typename' | 'author' | 'tag' | 'co
   tag?: Omit<Tag, '__typename' | 'pageCategory' | 'blogPosts'> & {
     pageCategory?: Omit<PageCategory, '__typename' | 'pages'>
   }
+}
+
+export type VznMeili = Omit<
+  Vzn,
+  '__typename' | 'mainDocument' | 'consolidatedText' | 'cancelationDocument' | 'amendmentDocument'
+> & {
+  mainDocument?: UploadFile
+  consolidatedText?: UploadFile
 }
