+++ next/backend/meili/utils.ts
@@ -1,18 +1,6 @@
 import { SearchResponse } from 'meilisearch'
 
-/**
- * A type that describes an entity wrapped in shared search index.
- * E.g.:
- * {
- *     type: "article",
- *     article: {...}
- * }
- */
-export type SearchIndexWrapped<T extends string, K extends object> = {
-  [key in T]: K
-} & {
-  type: T
-}
+import { SearchIndexWrapped } from './types'
 
 /**
  * Unwraps entities in Meilisearch response from shared search index.
@@ -23,3 +11,8 @@ export const unwrapFromSearchIndex = <T extends string, K extends object>(type: 
     return { ...response, hits: response.hits.map((hit) => hit[type] as K) }
   }
 }
+
+export const getMeilisearchPageOptions = ({ page, pageSize }: { page: number; pageSize: number }) => ({
+  limit: pageSize,
+  offset: (page - 1) * pageSize,
+})
