+++ next/components/molecules/SearchPage/BlogPostsResults.tsx
@@ -1,5 +1,5 @@
 // @ts-strict-ignore
-import { blogPostsFetcher, BlogPostsFilters, getBlogPostsSwrKey } from '@backend/utils/fetchers/blogPostsFetcher'
+import { blogPostsFetcher, BlogPostsFilters, getBlogPostsSwrKey } from '@backend/meili/fetchers/blogPostsFetcher'
 import { BlogItem, BlogSearchCards, LoadingSpinner } from '@bratislava/ui-bratislava'
 import useGetSwrExtras from '@utils/useGetSwrExtras'
 import { SearchResponse } from 'meilisearch'
