+++ next/components/molecules/sections/documentList/index.tsx
@@ -3,10 +3,10 @@
 // import { DocumentListFragment } from '@bratislava/strapi-sdk-homepage'
 import DocumentDevider from '@assets/images/documentDevider.svg'
 import DocumentDeviderSmall from '@assets/images/documentDevider-small.svg'
+import { MEILI_PAGE_SIZE, searchVZN } from '@backend/meili/meiliClient'
 import { BasicSearch, DocumentListItem, Modal, NoResultsFound, Pagination } from '@bratislava/ui-bratislava'
 import DocumentListCategorysMap from '@utils/documentListCategory'
 import { fileCountVzns } from '@utils/utils'
-import { MEILI_PAGE_SIZE, searchVZN } from 'backend/utils/meili'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
 import useSWR from 'swr'
