--- next/utils/getMeilisearchPageOptions.ts
@@ -1,4 +1,0 @@
-export const getMeilisearchPageOptions = ({ page, pageSize }: { page: number; pageSize: number }) => ({
-  limit: pageSize,
-  offset: (page - 1) * pageSize,
-})
